+++ app/src/main/java/com/superwall/superapp/MainApplication.kt
@@ -38,6 +38,7 @@ class MainApplication :
 
     override fun onCreate() {
         super.onCreate()
+        Superwall.instance.subscriptionStatus.value
         StrictMode.setThreadPolicy(
             ThreadPolicy
                 .Builder()