+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/get_presentation_result/GetPresentationResultLogic.kt
@@ -6,10 +6,10 @@ import com.superwall.sdk.paywall.presentation.result.PresentationResult
 object GetPresentationResultLogic {
     fun convertTriggerResult(triggerResult: InternalTriggerResult): PresentationResult =
         when (triggerResult) {
-            is InternalTriggerResult.EventNotFound -> PresentationResult.EventNotFound()
+            is InternalTriggerResult.PlacementNotFound -> PresentationResult.PlacementNotFound()
             is InternalTriggerResult.Holdout -> PresentationResult.Holdout(triggerResult.experiment)
             is InternalTriggerResult.Error -> PresentationResult.PaywallNotAvailable()
-            is InternalTriggerResult.NoRuleMatch -> PresentationResult.NoRuleMatch()
+            is InternalTriggerResult.NoAudienceMatch -> PresentationResult.NoAudienceMatch()
             is InternalTriggerResult.Paywall -> PresentationResult.Paywall(triggerResult.experiment)
         }
 }